--- eventmesh-admin-kotlin/src/main/java/com/apache/eventmesh/adminkotlin/EventmeshAdminApplication.java
@@ -1,13 +1,0 @@
-package com.apache.eventmesh.adminkotlin;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class EventmeshAdminApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(EventmeshAdminApplication.class, args);
-    }
-
-}
