--- eventmesh-admin-kotlin/src/main/java/com/apache/eventmesh/adminkotlin/dto/CommonResponse.java
@@ -1,22 +1,0 @@
-package com.apache.eventmesh.adminkotlin.dto;
-
-import lombok.Data;
-
-@Data
-public class CommonResponse {
-
-    private String data;
-
-    private String message;
-
-    public CommonResponse(String data) {
-        this.data = data;
-    }
-
-    public CommonResponse(String message, Exception e) {
-        this.message = message;
-        if (e != null) {
-            this.message += ": " + e.getMessage();
-        }
-    }
-}
