--- eventmesh-admin-kotlin/src/main/java/com/apache/eventmesh/adminkotlin/model/ConnectionInfo.java
@@ -1,5 +1,0 @@
-package com.apache.eventmesh.adminkotlin.model;
-
-public class ConnectionInfo {
-
-}
