--- eventmesh-admin-kotlin/src/main/java/com/apache/eventmesh/adminkotlin/model/SubscriptionInfo.java
@@ -1,4 +1,0 @@
-package com.apache.eventmesh.adminkotlin.model;
-
-public class SubscriptionInfo {
-}
