--- eventmesh-admin-kotlin/src/main/java/com/apache/eventmesh/adminkotlin/service/ConnectionService.java
@@ -1,5 +1,0 @@
-package com.apache.eventmesh.adminkotlin.service;
-
-public interface ConnectionService {
-
-}
