--- eventmesh-admin-kotlin/src/main/java/com/apache/eventmesh/adminkotlin/service/impl/EtcdConnectionService.java
@@ -1,13 +1,0 @@
-package com.apache.eventmesh.adminkotlin.service.impl;
-
-import org.springframework.stereotype.Service;
-
-import com.apache.eventmesh.adminkotlin.service.ConnectionService;
-
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-@Service
-public class EtcdConnectionService implements ConnectionService {
-
-}
