+++ eventmesh-admin-kotlin/src/main/java/org/apache/eventmesh/adminkotlin/dto/CommonResponse.java
@@ -15,4 +15,25 @@
  * limitations under the License.
  */
 
-rootProject.name = 'eventmesh-admin-new'
+package org.apache.eventmesh.adminkotlin.dto;
+
+import lombok.Data;
+
+@Data
+public class CommonResponse {
+
+    private String data;
+
+    private String message;
+
+    public CommonResponse(String data) {
+        this.data = data;
+    }
+
+    public CommonResponse(String message, Exception e) {
+        this.message = message;
+        if (e != null) {
+            this.message += ": " + e.getMessage();
+        }
+    }
+}
