+++ eventmesh-admin-kotlin/src/main/java/org/apache/eventmesh/adminkotlin/model/ConnectionInfo.java
@@ -15,4 +15,8 @@
  * limitations under the License.
  */
 
-rootProject.name = 'eventmesh-admin-new'
+package org.apache.eventmesh.adminkotlin.model;
+
+public class ConnectionInfo {
+
+}
