+++ eventmesh-admin-kotlin/src/main/java/org/apache/eventmesh/adminkotlin/model/SubscriptionInfo.java
@@ -15,4 +15,7 @@
  * limitations under the License.
  */
 
-rootProject.name = 'eventmesh-admin-new'
+package org.apache.eventmesh.adminkotlin.model;
+
+public class SubscriptionInfo {
+}
