+++ eventmesh-admin-kotlin/src/main/java/org/apache/eventmesh/adminkotlin/service/SubscriptionService.java
@@ -15,4 +15,18 @@
  * limitations under the License.
  */
 
-rootProject.name = 'eventmesh-admin-new'
+package org.apache.eventmesh.adminkotlin.service;
+
+import org.apache.eventmesh.adminkotlin.dto.CommonResponse;
+
+/**
+ * "Subscription" refers to the traditional MQ producer-consumer topic subscription relationship,
+ * emphasizing the subscription relationship between EventMesh clients (including SDK and connectors) and topics,
+ * reported by the EventMesh runtime.
+ */
+public interface SubscriptionService {
+
+    CommonResponse retrieveConfig(String dataId, String group);
+
+    String retrieveConfigs(Integer page, Integer size, String dataId, String group);
+}
