+++ eventmesh-admin-kotlin/src/test/java/org/apache/eventmesh/adminkotlin/EventmeshAdminApplicationTests.java
@@ -15,4 +15,16 @@
  * limitations under the License.
  */
 
-rootProject.name = 'eventmesh-admin-new'
+package org.apache.eventmesh.adminkotlin;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EventmeshAdminApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
