--- eventmesh-admin-new/src/main/java/com/apache/eventmesh/adminnew/EventmeshAdminApplication.java
@@ -1,13 +1,0 @@
-package com.apache.eventmesh.adminnew;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class EventmeshAdminApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(EventmeshAdminApplication.class, args);
-    }
-
-}
