--- eventmesh-admin-new/src/test/java/com/apache/eventmesh/adminnew/EventmeshAdminApplicationTests.java
@@ -1,13 +1,0 @@
-package com.apache.eventmesh.adminnew;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EventmeshAdminApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
